fix inline mixed fraction creation

+++ input/input.js
@@ -24,7 +24,6 @@ export class InputHandler {
             case "r":
                 return this.createSpecial("ROOT", "\\sqrt[{", "}]{", "}\\,");
             case "a":
-                if (currentToken.type === "MIXEDFRAC" && currentToken.exp === "start") return;
                 this.addToken({type:"ABS", exp:"start", rep: "|{"});
                 this.addToken({type:"ABS", exp:"end", rep: "}|"});
                 this.moveCursor("left");
@@ -53,6 +52,7 @@ export class InputHandler {
     delete(direction) {
         if (direction == "left") this.cursorPosition = Math.max(0, this.cursorPosition-1);
         const deleteToken = this.getCursorToken("right");
+        if (!deleteToken) return;
         if (deleteToken.type == "POWER") return this.deletePower();
         if (deleteToken && ["FRACTION", "MIXEDFRAC", "SQRT", "ROOT", "ABS"].includes(deleteToken.type)) {
             if (deleteToken.exp == "start") {
@@ -115,22 +115,21 @@ export class InputHandler {
         
         this.cursorPosition = orig + 1;
         const rightDigit = this.toNonDigit("right");
+        if (type === "MIXEDFRAC") this.addToken({type:type, exp:"middle2", rep: middle2});
         this.addToken({type:type, exp:"end", rep:end});
         
         // Position in the bottom of the fraction...
         this.cursorPosition = orig + 1;
         if (middle !== null) this.addToken({type:type, exp:"middle", rep:middle});
-        if (type === "MIXEDFRAC") {
-            this.addToken({type:type, exp:"middle2", rep: middle2})
-        }
 
-        if (leftDigit == orig && type == "FRACTION")
+        if (leftDigit === orig && type === "FRACTION" || type === "MIXEDFRAC")
             moveToTop = true; // Flag to move to first box if empty
-        if (type === "POWER" || type == "SQRT") {
+        if (type === "POWER" || type === "SQRT") {
             if (leftDigit !== orig) moveToTop = true
         }
 
         if (moveToTop && type === "FRACTION") this.cursorPosition = orig+1;  // Unless the top box is empty.
+        if (moveToTop && type === "MIXEDFRAC") this.cursorPosition = orig+2;  // Unless the top box is empty.
         if (!moveToTop && type === "POWER") this.cursorPosition = orig;
     }
     
@@ -168,7 +167,8 @@ export class InputHandler {
         while(this.getCursorToken(step)) {
             const currentToken = this.getCursorToken(step);
             const middleExp = currentToken.type === "FRACTION" ? "middle" : "middle2"
-            if (currentToken.type === "FRACTION" || currentToken.type === "MIXEDFRAC" ) {
+            if (currentToken.type === "FRACTION" || (currentToken.type === "MIXEDFRAC")) {
+                if (currentToken.type === "MIXEDFRAC" && ((step === "right" && currentToken.exp === "middle" && currentHeight >= 0))) break;
                 if (currentToken.exp === middleExp && currentHeight <= 0) {
                     this.moveCursor(step);
                     return true;
@@ -219,6 +219,16 @@ export class InputHandler {
             const currentToken = this.getCursorToken(step);
             if (token.type === "LPAREN") currentBracket += increment; 
             if (token.type === "RPAREN") currentBracket -= increment; 
+
+            if (currentBracket < 0 &&
+                !(currentHeight.FRACTION > 0 ||
+                currentHeight.MIXEDFRAC > 0 ||
+                currentHeight.POWER > 0 ||
+                currentHeight.ROOT > 0 ||
+                currentHeight.SQRT > 0 ||
+                currentHeight.ABS > 0)
+            ) break;
+            
             if (["FRACTION", "MIXEDFRAC", "POWER", "SQRT", "ROOT", "ABS"].includes(token.type)) {
                 if (step === "left" && token.exp !== "end" && currentHeight[token.type] === 0) break; 
                 if (step === "right" && token.exp !== "start" && currentHeight[token.type] === 0) break;
